+++ src/components/MenuScreen/MenuScreen.types.ts
@@ -1,0 +1,6 @@
+import { RootStackParamList } from '../../../App';
+
+export type MenuOption = {
+  title: string;
+  pageToNavigateTo: keyof RootStackParamList;
+};
