+++ src/components/MenuScreen/subcomponents/OuterCircle/OuterCircle.style.ts
@@ -1,0 +1,24 @@
+import { StyleSheet, TextStyle, ViewStyle } from 'react-native';
+
+type OuterCircleStyle = {
+  outerCircleContainer: ViewStyle;
+  outerCircleText: TextStyle;
+};
+
+const styles = StyleSheet.create<OuterCircleStyle>({
+  outerCircleContainer: {
+    width: 100,
+    height: 100,
+    borderRadius: 50,
+    position: 'absolute',
+    backgroundColor: 'white',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  outerCircleText: {
+    fontFamily: 'RubikBubbles-Regular',
+    textAlign: 'center',
+  },
+});
+
+export default styles;
